+++ src/game/tests.rs
@@ -381,6 +381,7 @@ mod vintage_cube_enduring_innocence;
 mod vintage_cube_ephemerate;
 mod vintage_cube_ertai_resurrected;
 mod vintage_cube_eternal_witness;
+mod vintage_cube_everflowing_chalice;
 mod vintage_cube_exhume;
 mod vintage_cube_expedition_map;
 mod vintage_cube_exploration;